--- app/src/main/java/projetsi/models/BlockingQueue.java
@@ -1,5 +1,0 @@
-package projetsi.models;
-
-public class BlockingQueue<T> {
-
-}
